requirements.txt
streamlit
requests
pandas
numpy
matplotlib
vaderSentiment
google-generativeai>=0.7.0
google-genai>=0.3.0
